Add custom JSON Button
Provides a custom 'json' button for the DataTables Buttons extension (https://datatables.net/extensions/buttons/). When clicked, for browsers which support opening blob data urls (only tested on chrome and firefox) it opnes a new window displaying the [nicely formatted] JSON data.

It can be used like any of the built-in button types:

$('#myTable').DataTable( {
    buttons: [
        'copy', 'excel', 'pdf', 'json'
    ]
} );