Don't rely on private DataTables object
We use the settings.nTable object to get the DOM table element for
performance reasons, even though the structure of the settings object is
private and might change with future versions of DataTables.

This change adds a fallback to re-create the API object (and get the
table node that way) if the settings.nTable object does not exist.