Fix setting of country property from loaded JSON
When loading from a local JSON object (as opposed to making an AJAX
call), it is likely that the 'country' property is already set (it is if
the JSON comes from fetchkivajson.js). This change uses that property if it is set.

This fixes the problem where the 'country' property was being set to
"object" from data saved by fetchkivajson.js